test: Add more tests
Finished writing tests for the first method.
The second and last method did not get a proper test.
I would need to investigate a lot more time into sinon to mock
requests and this seems not to be time well spent. Should have
tried vue or similar and then have (hopefully) testable components.